Various fixes

+++ src/config_net.h
@@ -71,7 +71,7 @@ public:
   bool                                    m_UDPDoNotRouteEnabled;       // whether to enable SO_DONTROUTE for UDP sockets
 
   bool                                    m_AllowDownloads;             // allow map downloads or not
-  uint16_t                                m_DownloadTimeout;
+  int32_t                                 m_DownloadTimeout;
   std::set<std::string>                   m_MapRepositories;             // enabled map repositories
   uint8_t                                 m_AllowTransfers;             // map transfers mode
   uint32_t                                m_MaxDownloaders;             // maximum number of map downloaders at the same time
